+++ src/lib.rs
@@ -19,18 +19,28 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate core;
 
+#[macro_use]
+extern crate memoffset;
 #[cfg(feature = "serde")]
 extern crate serde;
 #[cfg(feature = "stable_deref_trait")]
 extern crate stable_deref_trait;
 
 mod arc;
-mod borrow;
-mod unique;
+mod arc_borrow;
+mod arc_handle;
+mod arc_union;
+mod header;
+mod thin_arc;
+mod unique_arc;
 
 pub use arc::*;
-pub use borrow::*;
-pub use unique::*;
+pub use arc_borrow::*;
+pub use arc_handle::*;
+pub use arc_union::*;
+pub use header::*;
+pub use thin_arc::*;
+pub use unique_arc::*;
 
 #[cfg(feature = "std")]
 use std::process::abort;